Java
Create a class MyCalculator which consists of a single method long power (int, int). This method takes two integers, n and p, as parameters and finds np. If either n or p is negative, then the method must throw an exception which says "n and p should not be negative". Also, if both n and p are zero, then the method must throw an exception which says " n and p should not be be zero" For example, -4 and -5 would result in  java.lang.Exception: n and p should not be negative. Write a Java program for the function power in class MyCalculator and return the appropriate result after the power operation or an appropriate exception as detailed above.